address block manager refactor
